@@ -29,7 +29,7 @@
     "intl": "~1.0.0-rc-4",
     "jquery": "^2.1.1",
     "object-assign": "^2.0.0",
-    "rc-calendar": "~3.6.0",
+    "rc-calendar": "~3.12.4",
     "rc-dialog": "~4.0.7",
     "react": "^0.13.0",
     "react-art": "^0.13.0",
@@ -51,10 +51,12 @@
   "devDependencies": {
     "babel": "~5.4.7",
     "babel-core": "~5.4.7",
+    "babel-eslint": "~4.0.5",
     "babel-loader": "~5.1.3",
     "css-loader": "^0.9.1",
-    "eslint": "~0.21.2",
-    "eslint-plugin-react": "~2.3.0",
+    "eslint": "~1.1.0",
+    "eslint-config-airbnb": "0.0.7",
+    "eslint-plugin-react": "~3.2.2",
     "expect.js": "^0.3.1",
     "extract-text-webpack-plugin": "^0.3.8",
     "file-loader": "^0.8.1",
