@@ -1,38 +1,48 @@
-'use strict';
-
-var React = require('react');
+const React = require('react');
 
 module.exports = React.createClass({
-  getDefaultProps: function () {
+  propTypes: {
+    count: React.PropTypes.number,
+  },
+  getDefaultProps() {
     return {
-      count: 0
+      count: 0,
+      disabled: true,
     };
   },
-  getInitialState: function () {
+  getInitialState() {
     return {
       count: this.props.count,
-      disabled: true
     };
   },
-  increase: function () {
+  componentDidMount() {
+    this.turnOn();
+  },
+  componentWillReceiveProps(nextProps) {
     this.setState({
-      count: this.state.count + 1
+      count: nextProps.count,
     });
   },
-  componentDidMount: function () {
-    this.setState({
-      disabled: false
-    });
-  },
-  render: function () {
+  render() {
     return (
       <div>
         <h2>count</h2>
+
         <div>
           <p>{this.state.count}</p>
           <button onClick={this.increase} disabled={this.state.disabled}>increase</button>
         </div>
       </div>
     );
-  }
+  },
+  turnOn() {
+    this.setState({
+      disabled: false,
+    });
+  },
+  increase() {
+    this.setState({
+      count: this.state.count + 1,
+    });
+  },
 });
