@@ -1,16 +1,16 @@
-'use strict';
 
-var AppDispatcher = require('./AppDispatcher');
+
+const AppDispatcher = require('./AppDispatcher');
 
 module.exports = {
-  change: function () {
-    setTimeout(function () {
+  change() {
+    setTimeout(() => {
       AppDispatcher.handleViewAction({
         user: {
-          name: 'changed ' + Date.now()
+          name: 'changed ' + Date.now(),
         },
-        actionType: 'update_user'
+        actionType: 'update_user',
       });
     }, 1000);
-  }
+  },
 };
