@@ -1,15 +1,15 @@
-'use strict';
-var container;
+
+let container;
 require('./async.css');
-var Dialog = require('rc-dialog');
+const Dialog = require('rc-dialog');
 require('rc-dialog/assets/bootstrap.css');
-var React = require('react');
-exports.show = function () {
+const React = require('react');
+exports.show = () => {
   if (!container) {
     container = document.createElement('div');
     document.body.appendChild(container);
   }
-  React.render(<Dialog style={{width:400}}>
+  React.render(<Dialog style={{width: 400}}>
     <div className="my-content">loaded</div>
   </Dialog>, container).show();
 };
